LayoutTests/canvas/philip/tests/2d.clearRect.nonfinite-expected.txt

< [index] >
2d.clearRect.nonfinite
clearRect() with Infinity/NaN is ignored
References: 2d.nonfinite
Actual output:
Expected output:

Passed
